requirements.txt
PyYAML>=6.0
pytest>=7.0
pydantic>=1.10
